drivers/sensor: lsm6dso: Disable sensor hub before reset
Per an ST app note, the sensor hub I2C controller should be disabled
before doing a software reset.  Possibly, this is because the sensor hub
could be in the middle of the an I2C transaction to a sensor when it is
reset.  Disabling it and then waiting makes sure it has quiesced before
resetting.

(cherry picked from commit 25579d95b129c0eace4e0dbbffd936175648a556)

Original-Signed-off-by: Trent Piepho <trent.piepho@igorinstitute.com>
GitOrigin-RevId: 25579d95b129c0eace4e0dbbffd936175648a556
Change-Id: I9fae0ab8e18557a4e8fdaaa0cead55b2b0f853df
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4191692
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>