drivers/sensor: lsm6dso: Set to user bank on init
The lsm6dso initialization will fail if the device is not already set to
the user register bank.  All the registers used will be the wrong ones
from whatever bank it is in, e.g. sensor hub bank.  This includes the
registers to reset the device!

The bank will default to the user bank on reset, but the chip has no
hardware reset line.  On a reboot it will be in whatever bank it was
last in.  If the sensor hub is enabled, it will switch banks on every
sample, so it's entirely possible to reset or reboot when it happens
to be set to the sensor hub bank, which will cause the driver to
fail to initialize.  It will not work again until the lsm6dso is power
cycled.

(cherry picked from commit c379f86dbd9de887c4b673565eb844e7b835ff43)

Original-Signed-off-by: Trent Piepho <trent.piepho@igorinstitute.com>
GitOrigin-RevId: c379f86dbd9de887c4b673565eb844e7b835ff43
Change-Id: I9d15a0f38a4a3599bae08c544ffdcd2641d43138
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4191693
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>