drivers/sensor: lsm6dso: Remove busy wait on bank change
It's not necessary to busy wait 150 µs after changing register banks.
Nothing in the data sheet nor app note suggests this.  ST's own HAL,
which is used by this driver, does not delay when changing banks.  It
does a bank change around every function that accesses a non-user bank
register (it's quite inefficient).

So if it was necessary it would be broken now, as most of the bank
changes have no delay.

One of the few page changes that did have this delay are the those done
before and after reading a sensor sample.  Which is where the speed is
significant and is limiting the update rate the driver is capable of.

(cherry picked from commit c45bc68e5c8f633fd637659fd0c330bdf1f170db)

Original-Signed-off-by: Trent Piepho <trent.piepho@igorinstitute.com>
GitOrigin-RevId: c45bc68e5c8f633fd637659fd0c330bdf1f170db
Change-Id: Ib35a91191a2ec003704921929ce07cfea45ecee8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4191696
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>