modules: hal_nordic: nrf_802154: RNG xor-shift algorithm
The LCG method used earlier in the random number generator was problematic,
as the lowest bits repeated periodically (for example, the generated number
always resulted in an odd-even-odd-even-.. sequence, or the last three bits
formed an cycle of the length 8). This is because the LCG was done module
2^32. Any LCG using a power of 2 as the modulus will cause the same issue.
The used RNG method was changed to Marsaglia's xor shift-algorithm,
which does not have this issue.

(cherry picked from commit 4400540d15e2aa882c3c20f75ccafa32375e6a43)

Original-Signed-off-by: Artur Hadasz <artur.hadasz@nordicsemi.no>
GitOrigin-RevId: 4400540d15e2aa882c3c20f75ccafa32375e6a43
Change-Id: Idacd0b746c01a3a19d182c2c7166b96283bb11ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4138330
Commit-Queue: Keith Short <keithshort@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>