drivers: nrf_rtc_timer: Always announce actual number of elapsed ticks
In non-tickless mode, the timeout handler always announced maximum 1
tick to kernel, but in fact it cannot be guaranteed that the handler
execution is not delayed and that the number of elapsed ticks does not
exceed 1. Use the actual number instead.
Switch also to using a 32-bit value for `dticks` to get a bit simpler
generated code (ticks delta is not supposed to be that huge).

(cherry picked from commit 3ffaaa989a99e93202d2bfc6603f642e86ce9236)

Original-Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>
GitOrigin-RevId: 3ffaaa989a99e93202d2bfc6603f642e86ce9236
Change-Id: Ic7f0bf03ba1e78f49f6ddaa43570c23c14f5c8d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4176967
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>