drivers: nrf_rtc_timer: Correct initial timeout value
Values to be set to the comparator need to be specified in RTC cycles,
not ticks, so the initial value used in the tickless mode needs to be
MAX_CYCLES, otherwise when CONFIG_SYS_CLOCK_TICKS_PER_SEC is set to
a value less then the RTC frequency, the initially configured timeout
will be unnecessarily shorter.
On the occassion, remove also the call to counter() when setting the
initial timeout value in non-tickless mode. RTC is cleared a few lines
above, so at this point it will most likely be 0, and even if it was
not, compare_set() would properly handle a target time value that had
already passed.

(cherry picked from commit eb0cbb433c7471c8f1a10e91c04d46d912525993)

Original-Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>
GitOrigin-RevId: eb0cbb433c7471c8f1a10e91c04d46d912525993
Change-Id: Ia6df2e880311ae55b423d3b2a4b460d6903ff70a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4176968
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>