drivers: nrf_rtc_timer: Remove unnecessary interrupt locking
There is no need to disable interrupts while just checking if
a channel needs to be processed in the ISR, as that section
does not contain anything that needs to be protected against
overwriting from some other context. In particular, if a given
timeout is changed or even aborted while its event is being
checked, this will be correctly handled in the code that follows
and that checks the expiration time.

(cherry picked from commit 3454a2547ead7f867c75cf83b12a38a6ea4dc599)

Original-Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>
GitOrigin-RevId: 3454a2547ead7f867c75cf83b12a38a6ea4dc599
Change-Id: I26ee251fbb4c6e46bf6bec720bdca0cd66bee923
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4176969
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>