mm: intel_adsp_mtl_tlb: suppress sparse address space warnings
There are various call to z_soc_cached_ptr() which returns values
in the cached address space and are stored in temporary variables
that are not marked as in the cached address space. This results
in sparse complaining about discarding the cached address space
attribute. These temporary variables are then passed to other
internal memory management related functions which do not have
the concept of cached address space (as it is currently Xtensa
specific). Because of this, we cannot change the signature of
these functions. Instead, we force a change of address space
when those temporary variables are being assigned to suppress
sparse warnings.

(cherry picked from commit 422250d3b183a57982125a324c7382a18b540665)

Original-Signed-off-by: Daniel Leung <daniel.leung@intel.com>
GitOrigin-RevId: 422250d3b183a57982125a324c7382a18b540665
Change-Id: I48a797d0bf9b408705cf2909d4699368d532011e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4182068
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>