doc: application: improve example-application section
The 'Example Application' text is similar to but not the same as the
name of the repository, which is example-application. Use the name of
the repository instead. This is easier to search for and plants seeds
in people's memory about where to find it on GitHub.

Add more cross references.

(cherry picked from commit 873b63109b683dc45e539cc67cf9218314beb55f)

Original-Signed-off-by: Marti Bolivar <marti.bolivar@nordicsemi.no>
GitOrigin-RevId: 873b63109b683dc45e539cc67cf9218314beb55f
Change-Id: Ibf776b8df5d1fa1993ba5ae699b1fdf3f815390e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4147370
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Tristan Honscheid <honscheid@google.com>