samples: unicast_audio_client: Fix enabling the streams
This fixes the application that was not waiting for QoS configuration of
all streams in the group. The QoS operation is done on group of streams,
while the status of each one of the streams is received separately.
Meaning that the application has to control which streams were already
configured before enabling them.

(cherry picked from commit 18ab7ebf220fda158a9b4437aa4cebd26c686ae5)

Original-Signed-off-by: Mariusz Skamra <mariusz.skamra@codecoup.pl>
GitOrigin-RevId: 18ab7ebf220fda158a9b4437aa4cebd26c686ae5
Change-Id: I26b485e70d3edbb92afc40fa723a91c55c375949
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4177001
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>