west: runners: blackmagicprobe: update reset on connect command
Recently the blackmagicprobe command for "connect under reset" has been
updated from "connect_srst" to "connect_rst". Update the flag name in
the west blackmagicprobe binding, keep the old one as well for
compatibility with out of tree boards, and send both the old and new
commands to GDB so that it works with both the old and new firmware. GDB
is going to get a "command not supported by this target" for one of the
two and proceed with the other one.

(cherry picked from commit dabfffaf26afa8cf03f00e4fc4306de0a6e70375)

Original-Signed-off-by: Fabio Baltieri <fabio.baltieri@gmail.com>
GitOrigin-RevId: dabfffaf26afa8cf03f00e4fc4306de0a6e70375
Change-Id: I8f4fb1125c1110d9b1de4e2422cef5431a86f954
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4180796
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>