tests/disk: Add NVMe disk to disk_performance
On Qemu x86_64 it is possible to emulate an NVMe controller and disk on
the PCIe bus, so let's add this case here to verify how the newly added
NVMe controller and disk driver perform, relatively to the host.
This will prove to be useful to evaluate the optimization to be made, or
the possible regression that could happen on future changes.

(cherry picked from commit 8a322cb56d37560ac9e2e21f187e763ec21a2828)

Original-Signed-off-by: Tomasz Bursztyka <tomasz.bursztyka@linux.intel.com>
GitOrigin-RevId: 8a322cb56d37560ac9e2e21f187e763ec21a2828
Change-Id: I7ebff93ba1cae53469c523f40f1a30534d8cf5d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4410696
Commit-Queue: Yuval Peress <peress@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Yuval Peress <peress@google.com>
Reviewed-by: Yuval Peress <peress@google.com>