arch: arm64: Refine v8R AArch64 MPU regions switch
The current mechanism of the MPU region switching configures and
reprograms the regions (including inserting, splitting the dynamic
region, and flushing the regions to the registers) every time during the
context switch. This, not only causes a large usage of the kernel stack
but also a lower performance.

To improve it, move the configuration operations ahead to make sure the
context swtich only flushes the current thread regions to the registers
and does not configure the regions anymore. To achieve this, configure
the regions during any operations related to partitions (partition
add/remove, and domain add/remove thread), flush the sys_dyn_regions if
the current thread is the privileged thread, and flush the thread's own
regions if it's a user thread.

(cherry picked from commit 0f7bbff050fe8fe334040befc6f2ef7f27e79210)

Original-Signed-off-by: Jaxson Han <jaxson.han@arm.com>
GitOrigin-RevId: 0f7bbff050fe8fe334040befc6f2ef7f27e79210
Change-Id: I51dae906def6371736bb8a6d5c62bb99b81a72bb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4797916
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>