Revert "Bluetooth: Host: Disable host-based RPA resolution when !BT_PRIVACY"
This reverts commit 5824ac90ecd8dd06d9ea117d345dbc6b069e5840.

The resulting revert takes into account changes to the predicate that
guards the use of `resp_addr`.

In the resulting revert, the predicate guarding the use of `resp_addr`
is differs from the naive revert because the intention of that predicate
is to know when `conn.c:bt_conn_le_create` decided to use the Host
resolver. And, the logic in `conn.c:bt_conn_le_create` has changed. The
changes to the logic in `conn.c` are in the following commits:

89780d715d72cca55101f66965c1ac05024b45cc
5a09325ba966e75b9ac8e176298b54fb8a655aee

(cherry picked from commit 4b675f34dbbaab1f81743a764c8d8019cdb00cac)

Original-Signed-off-by: Aleksander Wasaznik <aleksander.wasaznik@nordicsemi.no>
GitOrigin-RevId: 4b675f34dbbaab1f81743a764c8d8019cdb00cac
Change-Id: Iad1dd7db48910ea8509542432d36bcf102f27292
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4194352
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>