Revert "twister: Update path structure for tests"
This reverts commit 21eb27c5c0ef66ca98a1a874fcdb5e29226a33d6.

This change has been a source of much trouble and ends up preventing us
from tracking test results across tree. It uses the repo name as the
namespace, and that is not always the same and does not have to be
called 'zephyr', depending on where you are running and in which
environment.

(cherry picked from commit a4b2d58d3f5dc08c403990133eeee052347ef51e)

Original-Signed-off-by: Anas Nashif <anas.nashif@intel.com>
GitOrigin-RevId: a4b2d58d3f5dc08c403990133eeee052347ef51e
Change-Id: I8a4593c7343026d53e7268232357b84f4dac2a6a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4194381
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>