drivers: gpio_rpi_pico.c: add support for single-ended IO
register to a constant value, and then changing the GPIO_OE register
instead, when the output has to change. To do this, the output-setting
functions need to know which pins have been configured as single-ended,
and for that reason the data structure has been extended to include
this information.

Another change is the PR, is that setting of the pull-ups/pull-downs
now applies to both inputs and outputs as well. Previous solution was
wrong, because if the user wanted to set up an input with a pull
resistor enabled, and then reconfigure it to an output without any
pulls, then the pulls remained in place for the output. Now pulls
are correctly set based on the gpio flags for outputs too, and this
is especially useful for single-ended outputs too.

(cherry picked from commit 38f554ef4f99281fc9d546da8d13eefccc03d30c)

Original-Note: RP2040 can support single-ended IO, by setting the GPIO_OUT
Original-Signed-off-by: Purdea Andrei <andrei@purdea.ro>
GitOrigin-RevId: 38f554ef4f99281fc9d546da8d13eefccc03d30c
Change-Id: Iad7ddc623b75065c8b710f24ee58f494a680a1bd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4291373
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>