mgmt: mcumgr: Add command result codes (ret)
Adds a new map to responses that command handlers can use to
return result codes of functions themselves, unrelated to SMP
error codes. This is achieved by adding an optional "ret" map
which contains a 16-bit "group" and 16-bit "rc" error code,
this map will not be included if the "ret rc" is 0 (which
implies success).

In order to allow this, a new version of the protocol has been
which can be indicated for use by setting a bit in the header.

(cherry picked from commit 6f5b9749499156a4ea8181aa5d32d447ecc7afc7)

Original-Signed-off-by: Jamie McCrae <jamie.mccrae@nordicsemi.no>
GitOrigin-RevId: 6f5b9749499156a4ea8181aa5d32d447ecc7afc7
Change-Id: Ic7b7c6abf6f1a2fac0d1b65de3715adb443f4700
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4532241
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Yuval Peress <peress@google.com>
Commit-Queue: Yuval Peress <peress@google.com>
Tested-by: Yuval Peress <peress@google.com>