bmi160: move emulator to driver directory
Allow the emulator to sit in the same directory as the driver
implementation. This will make working on the emulator much easier and
keep the emulator subsystem directory clean by allowing it to focus on
the actual subsystem and not the use cases of the subsystem.

(cherry picked from commit a6326f1f7fa285e64bcae43b67bdf05cc00355f9)

Original-Signed-off-by: Yuval Peress <peress@google.com>
GitOrigin-RevId: a6326f1f7fa285e64bcae43b67bdf05cc00355f9
Change-Id: Iba3654695fd5c2f8c30ebee0c8bc5b68a4c2a383
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4291298
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>